Add files via upload
This code integrates the camera functionality with the rest of the app. Now all we need to do is to add a button to analyse the pictures taken from the camera button! The issue I was running into I think was related to calling the camera file separately and this somehow caused the button not to work when I ran it. Remember this for the future!